% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid.R
\name{add_grid}
\alias{add_grid}
\title{Add grid}
\usage{
add_grid(zoo, grid)
}
\arguments{
\item{zoo}{zoo monitor tibble}

\item{grid}{grid object}
}
\value{
zoo grid object
}
\description{
Add grid
}
\examples{
pacman::p_load(tidyverse, targets)
data(skink)
grid44 <- create_grid(range(skink$X), range(skink$Y), dim = c(4,4))
grid44

add_grid(skink, grid44)
}
