% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-neigh-mean.R
\name{add_neigh_mean}
\alias{add_neigh_mean}
\title{add_neigh_mean}
\usage{
add_neigh_mean(x, grid_obj)
}
\arguments{
\item{x}{vector of observations}

\item{grid_obj}{grid object}
}
\value{
vector of neighbour means
}
\description{
Get average of neighbour observations
}
\examples{
grid_example <- grid
count_grid_behaviour(grid_example, skink_grid) |> 
dplyr::mutate(
 neigh_mean = add_neigh_mean(obs, grid_example)
)
}
