% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdNET_process}
\alias{birdNET_process}
\title{Runs \code{birdnetR::predict.birdnet_model_acoustic} with custom settings}
\usage{
birdNET_process(
  audio,
  model = c("BirdNET v2.4", "Perch v2"),
  slist = NULL,
  language = "de",
  batch_size = NULL,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25
)
}
\arguments{
\item{audio}{Audio file to process.}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}

\item{slist}{Optional. Character vector or list of species names to filter the predictions}

\item{language}{Character. defaults to \code{'de'}}

\item{batch_size}{Optional. An integer specifying the number of audio segments evaluated per inference call.}

\item{min_confidence}{Numeric value to set the minimum confidence threshold for predictions.}

\item{chunk_overlap_s}{Numeric value to set the overlap between consecutive segments.}

\item{sigmoid_sensitivity}{Numeric value that adjusts the sensitivity of the sigmoid function. Must be in the interval [0.5, 1.5]. defaults to \code{1.25}}
}
\value{
An S3 object of class \code{birdnet_prediction_acoustic} and \code{birdnet_prediction} containing the prediction results
}
\description{
Runs \code{birdnetR::predict.birdnet_model_acoustic} with custom settings on audio file.
}
\seealso{
\code{\link[birdnetR:predict.birdnet_model_acoustic]{birdnet.predict()}}
}
