% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{birdNET_process_batch}
\alias{birdNET_process_batch}
\title{Split execution of \link{birdNET_process} in sets of wave files}
\usage{
birdNET_process_batch(
  wave_files,
  model = c("BirdNET v2.4", "Perch v2"),
  n = 12,
  slist = NULL,
  language = "de",
  batch_size = 3,
  restart = NULL,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25
)
}
\arguments{
\item{wave_files}{audio files to process}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}

\item{n}{number of audio files to process in a single run}

\item{slist}{Optional. Character vector or list of species names to filter the predictions}

\item{language}{Character. defaults to \code{'de'}}

\item{batch_size}{Optional. An integer specifying the number of audio segments evaluated per inference call.}

\item{restart}{optional. restart iteration from here}

\item{min_confidence}{Numeric value to set the minimum confidence threshold for predictions.}

\item{chunk_overlap_s}{Numeric value to set the overlap between consecutive segments.}

\item{sigmoid_sensitivity}{Numeric value that adjusts the sensitivity of the sigmoid function. Must be in the interval [0.5, 1.5]. defaults to \code{1.25}}
}
\description{
Split execution of \link{birdNET_process} in sets of wave files
}
