% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdNET_select}
\alias{birdNET_select}
\title{Filter BirdNET selection tables by species}
\usage{
birdNET_select(
  path = NULL,
  model = c("BirdNET v2.4", "Perch v2"),
  slist = system.file("extdata/ornitho_de.txt", package = "MonitoR")
)
}
\arguments{
\item{path}{path to results}

\item{model}{one of 'BirdNET v2.4' or 'Perch v2'}

\item{slist}{path to custom species list}
}
\description{
Filter BirdNET selection tables by species
}
