% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ei.R
\name{calc_ei}
\alias{calc_ei}
\title{calc ei}
\usage{
calc_ei(df)
}
\arguments{
\item{df}{tibble with columns zone, obs and n of grid for zone}
}
\value{
tibble with calculation and ei
}
\description{
calc ei
}
\examples{
df1 <- tibble::tibble(
  zone = 1:2,
  obs = c(0, 100),
  n_grids = rep(1,2)
)
df2 <- tibble::tibble(
  zone = 1:2,
  obs = c(50,50),
  n_grids = rep(1,2)
)
df2
calc_ei(df1) |> print()
calc_ei(df2) |> print()
}
