% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spi.R
\name{calc_spi}
\alias{calc_spi}
\title{calculate SPI}
\usage{
calc_spi(zone_obj)
}
\arguments{
\item{zone_obj}{zone obj}
}
\value{
SPI
}
\description{
calculate SPI
}
\examples{
data(skink_grid)
data(grid)
grid$zone[1:2] <- 2
grid
get_zone_object(grid, skink_grid) |> calc_spi()
}
