% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canning_surfer.R
\name{canning_surfR}
\alias{canning_surfR}
\title{Reads in external data, combines with internal datasets and produces Canning
River surfer plots}
\usage{
canning_surfR(path)
}
\arguments{
\item{path}{Character string filepath to location of sonde data xlsx
workbooks.}
}
\value{
A pdf format four panel surfer plot of the Canning River.
}
\description{
\code{canning_surfR} takes a file path to Canning River sonde output and creates
    a four panel (single column) surfer plot of salinity, dissolved oxygen,
    chlorophyll a, and temperature in pdf format.
}
\details{
Surfer plots display a
    cross-section of the river where the metrics of interest have been
    interpolated between sonde locations. Inverse distance weighting has
    been used for the interpolation.

A river "bottom" is displayed which puts the interpolation in context.
    The river bottom has been derived from a combination of historical
    maximum depths at sampling locations as well as station points extracted
    from a "best" navigation line over the latest corporate bathymetry. See
    project documentation for further details.
}
\examples{
\dontrun{
canning_surfR(path = "Z:/DEC/MonitoringProgram/Data")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
