% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cassowary}
\alias{cassowary}
\title{Zoo monitor data for cassowary dataset}
\format{
A data frame with 7261 rows and 4 variables:
\describe{
\item{time}{date and time of observation}
\item{X}{x position}
\item{Y}{y position}
\item{behaviour}{observed behaviour}
\item{cassowary}{name of cassowary}
\item{jeffery_side}{side of jeffery}
\item{martina_side}{side of martina}
}
}
\source{
Eduardo Fernandez
}
\usage{
cassowary
}
\description{
A dataset containing cassowary data
}
\keyword{datasets}
