% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoo.R
\name{clean_zoo}
\alias{clean_zoo}
\title{clean_zoo}
\usage{
clean_zoo(zoo, beh_col = "none")
}
\arguments{
\item{zoo}{zoo monitor tibble}

\item{beh_col}{name of column with behaviour info}
}
\value{
tibble with time, X, Y and behaviour
}
\description{
clean_zoo
}
