% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_grid_behaviour.R
\name{count_grid_behaviour}
\alias{count_grid_behaviour}
\title{count_grid behaviour}
\usage{
count_grid_behaviour(grid, zoo_grid)
}
\arguments{
\item{grid}{grid object}

\item{zoo_grid}{zoo grid object}
}
\value{
tibble with grid number and number of obs
}
\description{
For each grid counts the number of observations in that
grid. Have to pass in grid object so that can get
the empty grid cells.
}
\examples{
count_grid_behaviour(grid, skink_grid)
}
