\name{dateParsed}
\alias{dateParsed}
\alias{file.date.info}

\title{Utilities to handle small S/R differences}
\description{Utilities to handle small S/R differences.}

\usage{
    dateParsed()
    file.date.info(file)
}

\arguments{
    \item{file}{a character string indicating a file name.}
}
\value{misc.}
\details{
    Utilities to handle small S/R differences. 
}
\seealso{
    \code{\link{is.S}}
    \code{\link{Sys.info}}
}
\examples{dateParsed() }
\keyword{programming}
\keyword{utilities}


