% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_pkgs}
\alias{.check_pkgs}
\title{Interactively check, load and optionally install missing R packages}
\usage{
.check_pkgs(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names to check/install.}
}
\description{
Interactively check, load and optionally install missing R packages
}
