% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{export_visuals}
\alias{export_visuals}
\title{Export visuals}
\usage{
export_visuals(path, model = c("BirdNET v2.4", "Perch v2", "combined"))
}
\arguments{
\item{path}{path}

\item{model}{Model used to classify; one of \code{"BirdNET_V2.4"} or
\code{"Perch v2"}}
}
\description{
Export visuals
}
