% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entropy_null.R
\name{get_entropy_null}
\alias{get_entropy_null}
\title{get_entropy_null}
\usage{
get_entropy_null(obs_grid, n_sims, n_grid)
}
\arguments{
\item{obs_grid}{observed grid for each observation}

\item{n_sims}{the number of simulations to do}

\item{n_grid}{the number of grid cells}
}
\value{
a vector of entropy for each permutations
}
\description{
Get an empirical distribution of entropy for a given number of
grids assuming evenly spread
}
\examples{
get_entropy_null(1:16, n_sims = 10, n_grid = 16)
}
