% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure_summary.R
\name{get_measure_summary}
\alias{get_measure_summary}
\title{get measure summary}
\usage{
get_measure_summary(grid, zoo_grid)
}
\arguments{
\item{grid}{grid object}

\item{zoo_grid}{zoo grid object}
}
\value{
summary of zones
}
\description{
get measure summary
}
\examples{
data(grid)
grid$zone[2] <- 2
data(skink_grid)
get_measure_summary(grid, skink_grid)
}
