% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zone_object.R
\name{get_zone_object}
\alias{get_zone_object}
\title{get zone object}
\usage{
get_zone_object(grid, zoo_grid)
}
\arguments{
\item{grid}{grid object}

\item{zoo_grid}{zoo grid object}
}
\value{
zone object
}
\description{
get zone object
}
\examples{
data(grid)
grid$zone[2] <- 2
data(skink_grid)
get_zone_object(grid, skink_grid)
}
