% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid}
\alias{grid}
\title{grid object for skink}
\format{
A data frame with 16 rows and 8 variables:
\describe{
\item{grid}{grid id}
\item{X}{x position}
\item{Y}{y position}
\item{left}{left edge}
\item{right}{right edge}
\item{bottom}{bottom edge}
\item{top}{top edge}
\item{zone}{zone}
}
}
\usage{
grid
}
\description{
A dataset containing wine reviews.
}
\keyword{datasets}
