% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_WTs_func.R
\name{identify_WTs_func}
\alias{identify_WTs_func}
\title{Identifies the wind turbines if several unit IDs appear.}
\usage{
identify_WTs_func(dataframe, id = names(dataframe[1]))
}
\arguments{
\item{dataframe}{A data frame containing the data from condition monitoring.}

\item{id}{The name as a character of the column with unit IDs.}
}
\value{
The data in \code{dataframe} sorted by \code{id} into a list containing dataframes, one for each wind turbines, and the time stamps are converted to POSIXct class.
}
\description{
Identifies the wind turbines if there are several unit IDs, and returns the same data in a list containing dataframes for each unit ID, and the time stamps are converted to POXIXct class.
}
\examples{
identify_WTs_func(data, id = "UnitID")
}

