% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_func_2mins.R
\name{interpolate_func_2mins}
\alias{interpolate_func_2mins}
\title{Interpolation function using 2 minute intervals.}
\usage{
interpolate_func_2mins(dataframe, var = "PowerActual")
}
\arguments{
\item{dataframe}{A dataframe containing a single wind turbine case.}

\item{var}{The variable to be interpolated in \code{dataframe}.}
}
\value{
\item{out}{A data frame with the new time stamps in the first column and the second column
contains interpolated \code{var} using 2 minute intervals.}
}
\description{
The interpolation of a single wind turbine with an interval length of 2 minutes. If there
are intervals larger than 10 minuttes with missing data no interpolation is done and NAs are added.
}
\examples{
interpolate_func_2mins(data, var = "GeneratorSpeed")

}

