% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_func_2mins_identified.R
\name{interpolate_func_2mins_identified}
\alias{interpolate_func_2mins_identified}
\title{Interpolation of identified data frame using 2 minute intervals.}
\usage{
interpolate_func_2mins_identified(dataframes, var = "PowerActual")
}
\arguments{
\item{dataframes}{A list of data frames as a result of the identified function with several
wind turbines.}

\item{var}{The variable to be interpolated in \code{dataframes}.}
}
\value{
\item{out}{A data frame with the new time stamps in the first column and the next columns,
one for each wind turbine,
contain interpolated \code{var} using 2 minute intervals.}
}
\description{
The interpolation of several wind turbines with an interval length of 2 minutes. If there
are intervals larger than 10 minuttes the missing data no interpolation is done and NAs are added.
}
\examples{
interpolate_func_2mins(data, var = "GeneratorSpeed")
}

