\name{is.S}
\alias{is.S}

\title{S/R System Information}
\description{Determine if the system is S or R.}

\usage{
    is.S()
}

\arguments{None}
\value{TRUE or FALSE}
\details{
    If the system is S return TRUE, otherwise FALSE. 
}
\seealso{
    \code{\link{is.R}}
}
\examples{is.R() }
\keyword{programming}
\keyword{utilities}

