% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_arma.R
\name{kalman_filter_arma}
\alias{kalman_filter_arma}
\title{Kalman filtering a univariate stationary zero-mean ARMA process.}
\usage{
kalman_filter_arma(ts, F, G, Q, m0, C0)
}
\arguments{
\item{ts}{A univariate time series with zero mean}

\item{F}{The coefficient matrix in the observation equation, as shown above.}

\item{G}{The matrix in the state equation as shown above.}

\item{Q}{The variance matrix of the state equation.}

\item{m0}{The initial value of \eqn{\bm{x}_t}.}

\item{C0}{The initial value of the state variance.}
}
\value{
The return is a list of the innovations, standardised residuals, and predicted values.
\item{innovations}{The innovations.}
\item{sd}{the standardised residuals.}
\item{y_predicted}{the predicted values.}
}
\description{
Function that applies the Kalman filter to a univariate stationary zero-mean ARMA process.
To get the Kalman filter to work the process should be written as a dynamic linear model.
The ARMA process written as a dynamic linear model has the form,
where \eqn{y_t = F\bm{x}_t} is the
observation equation and \eqn{\bm{x}_t = G\bm{x}_{t-1} + H w_t} is the state equation,
and \eqn{Q = \mbox{Var}(H w_t)}.
}
\examples{
kalman_filter_arma(ts = data, F=F, G=G, Q=Q, m0=m0, C0=C0)

}

