% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_random_walk.R
\name{Kalman_filter_random_walk}
\alias{Kalman_filter_random_walk}
\title{Kalman filter applied to multivariate random walk.}
\usage{
Kalman_filter_random_walk(ts, F = c(1, 1, 1, 1), R = 0.1 * diag(4),
  Q = 0.1, m0 = 0, C0 = 1)
}
\arguments{
\item{ts}{Multivariate time series}

\item{R}{The covariance of the meassurement noise.}

\item{Q}{The covariance of the state noise.}

\item{m0}{Initial state.}

\item{C0}{Initial covariance of the state process.}
}
\value{
State values and state covariances.
}
\description{
Kalman filter applied to multivariate random walk.
}
\examples{
Kalman_filter_random_walk(data)
}

