% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPCA.R
\name{KPCA_model}
\alias{KPCA_model}
\title{kernel PCA for process monitor}
\usage{
KPCA_model(data, kernel, kernel_num, ...)
}
\arguments{
\item{data}{A centered-and-scaled data matrix}

\item{kernel}{The kernel function to be used to calculate the kernel matrix. 
This has to be a function of class kernel,
i.e. which can be generated either one of the build in kernel generating 
functions (e.g., rbfdot etc.) or a user defined function of 
class kernel taking two vector arguments and returning a scalar.}

\item{kernel_num}{The number of principle component}

\item{...}{Lazy dots for additional internal arguments}
}
\value{
A list of class 'KPCA' with following
\itemize{
  \item{projectionMatrix -- }{the q eigenvectors corresponding to the q
    largest eigenvalues as a p x q projection matrix}
  \item{LambdaInv -- }{the diagonal matrix of inverse eigenvalues}
  \item{SPE -- }{the vector of SPE test statistic values for each of the n
    observations contained in "data"}
  \item{T2 -- }{the vector of Hotelling's T2 test statistic for each of the
    same n observations}
  \item{model_data -- }{the data used to build the model}
  \item{K_hat -- }{The mean and center matrix of kernel matrix}
  \item{eigenK -- }{A list consists of eigen values and eigen vector of K_hat}
}
}
\description{
Use kernel principle component analysis for process monitoring
and also find the squared prediction error (SPE) and Hotelling's T2 test
statistic values for each observation in this data matrix.
}
\references{
http://modeleau.fsg.ulaval.ca/fileadmin/modeleau/documents/Publications/pvr487.pdf
}
