\name{nseriesMethods}
\alias{nseriesOutput,IOseriesIDs-method}
\alias{nseriesInput,IOseriesIDs-method}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
    \S4method{nseriesOutput}{IOseriesIDs}(x)
    \S4method{nseriesInput}{IOseriesIDs}(x)
}
\arguments{
    \item{x}{a IOseriesIDs object.}
%%    \item{series}{series to select. Passed to selectSeries.}
}
%\keyword{DSE}
\keyword{ts}


