% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_monitor.R
\name{PCA_model}
\alias{PCA_model}
\title{PCA model for process monitor}
\usage{
PCA_model(data, kernel_num, ...)
}
\arguments{
\item{data}{A centered-and-scaled data matrix}

\item{kernel_num}{The number of principle component}

\item{...}{Lazy dots for additional internal arguments}
}
\value{
A list of class "pca" with the following:
\itemize{
  \item{projectionMatrix -- }{the q eigenvectors corresponding to the q
    largest eigenvalues as a p x q projection matrix}
  \item{LambdaInv -- }{the diagonal matrix of inverse eigenvalues}
  \item{SPE -- }{the vector of SPE test statistic values for each of the n
    observations contained in "data"}
  \item{T2 -- }{the vector of Hotelling's T2 test statistic for each of the
    same n observations}
}
}
\description{
Calculate the principal component analysis for process monitor,
and also find the squared prediction error (SPE) and Hotelling's T2 test
statistic values for each observation in this data matrix.
}
\examples{

}
\references{
https://github.com/gabrielodom/mvMonitoring
}
