% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_monitor.R
\name{plot.threshold}
\alias{plot.threshold}
\title{Plot for the class threshold}
\usage{
\method{plot}{threshold}(threshold_object)
}
\arguments{
\item{threshold_object}{The class of threshold}
}
\value{
A plot
}
\description{
Plot the statistics T2 and SPE of class threshold.
}
