% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entropy.R
\name{plot_entropy}
\alias{plot_entropy}
\title{plot entropy}
\usage{
plot_entropy(df, calc = "Grid")
}
\arguments{
\item{df}{tibble from get measure}

\item{calc}{whether to plot entropy based on grid or zones.}
}
\value{
plot
}
\description{
plot entropy
}
