% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{plot_grid}
\alias{plot_grid}
\title{plot_grid}
\usage{
plot_grid(grid, zoo, grid_col = FALSE, zone_fill = FALSE, label = TRUE)
}
\arguments{
\item{grid}{grid object}

\item{zoo}{zoo object}

\item{grid_col}{fill to show grid name}

\item{zone_fill}{show zone colour}

\item{label}{boolean to show grid labels}
}
\value{
plot of grid on zoo data
}
\description{
plot_grid
}
\examples{
data(skink)
grid <- create_grid(range(skink$X), range(skink$Y), dim = c(4,4))
grid$zone[1] <- 2
plot_grid(grid, skink, zone_fill = TRUE)
}
