% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neigh.R
\name{plot_neigh}
\alias{plot_neigh}
\title{plot_neigh}
\usage{
plot_neigh(grid_obj, cell)
}
\arguments{
\item{grid_obj}{grid object}

\item{cell}{cell number}
}
\value{
plot of neighbours
}
\description{
plot_neigh
}
\examples{
plot_neigh(grid, 6)
}
