\name{print.IOseriesIDs}
\alias{print.IOseriesIDs}
\title{Print Specific Methods}
\description{See the generic function description.}
\usage{
    \method{print}{IOseriesIDs}(x, ...)
    }
\arguments{
    \item{x}{a IOseriesIDs object.}
    \item{...}{arguments to be passed to other methods.}
}
\seealso{
    \code{\link{print}}
    \code{\link{summary}}
}
%\keyword{DSE}
\keyword{ts}

