% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_audacity}
\alias{read_audacity}
\title{read audacity labels}
\usage{
read_audacity(
  path,
  pattern = c("BirdNET.results.txt", "BirdNET.labels.txt"),
  delim = "\\t"
)
}
\arguments{
\item{path}{path}

\item{pattern}{pattern to recognise}

\item{delim}{delimeter}
}
\description{
read audacity labels
}
