% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_BirdNET_analysis_params}
\alias{read_BirdNET_analysis_params}
\title{Read BirdNET_analysis_params}
\usage{
read_BirdNET_analysis_params(path, pattern = "BirdNET_analysis_params.csv")
}
\arguments{
\item{path}{path}

\item{pattern}{defaults to 'BirdNET_analysis_params.csv'}
}
\description{
Read BirdNET_analysis_params
}
