% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMMICRO2.R
\name{read_params}
\alias{read_params}
\title{Read summary.txt}
\usage{
read_params(path, pattern = "Summary.txt")
}
\arguments{
\item{path}{path}

\item{pattern}{defaults to 'Summary.txt'}
}
\description{
Read summary.txt
}
