% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoo.R
\name{read_zoo}
\alias{read_zoo}
\title{read_zoo}
\usage{
read_zoo(file, sheet = 1)
}
\arguments{
\item{file}{zoo monitor file}

\item{sheet}{file sheet to use}
}
\value{
returns a tibble with all the data in file
}
\description{
read_zoo
}
