% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_photo2datetime}
\alias{rename_photo2datetime}
\title{Rename file names to YYYYMMDD_HHMMSS format}
\usage{
rename_photo2datetime(
  input_dir,
  output_dir = NULL,
  recursive = T,
  pattern = c("JPG", "JPEG", "PNG"),
  prefix = NULL,
  dtime = c("ctime", "mtime"),
  tz = ""
)
}
\arguments{
\item{input_dir}{input folder}

\item{output_dir}{output folder}

\item{recursive}{boolean}

\item{pattern}{file extension}

\item{prefix}{string}

\item{dtime}{'mtime' or 'ctime'}

\item{tz}{time zone}
}
\description{
Retrieve timestamps from file meta data and rename accordingly. If output_dir is specified files are copied, otherwise renamed.
}
