% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{run_birdnet}
\alias{run_birdnet}
\title{Split execution of \link{birdNET_process} in sets of wave files}
\usage{
run_birdnet(
  wave_files,
  model = c("BirdNET v2.4", "Perch v2"),
  slist = NULL,
  language = "de",
  batch_size = 3,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25,
  skip.existing.results = FALSE
)
}
\arguments{
\item{wave_files}{audio files to process}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}

\item{slist}{Optional. Character vector or list of species names to filter the predictions}

\item{language}{Character. defaults to \code{'de'}}

\item{batch_size}{Optional. An integer specifying the number of audio segments evaluated per inference call.}

\item{min_confidence}{Numeric value to set the minimum confidence threshold for predictions.}

\item{chunk_overlap_s}{Numeric value to set the overlap between consecutive segments.}

\item{sigmoid_sensitivity}{Numeric value that adjusts the sensitivity of the sigmoid function. Must be in the interval [0.5, 1.5]. defaults to \code{1.25}}

\item{skip.existing.results}{boolean}
}
\description{
Replaces \link{birdNET_process_batch}
}
