% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_pca.R
\name{scale_pca}
\alias{scale_pca}
\title{Scale function for PCA model}
\usage{
scale_pca(normal, fault, ...)
}
\arguments{
\item{normal}{The standard matrix is used to scale}

\item{fault}{The matrix is used to be scaled. W}

\item{...}{Lazy dots for additional internal arguments}
}
\value{
A matrix of the scaled matrix fault
}
\description{
Before the data is used for PCA, the data should be centered and scaled
}
\examples{
data(normal_data)
normal_scale <- scale_pca(normal_data)
}
