% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skink_grid}
\alias{skink_grid}
\title{Zoo monitor data for skink with added grid}
\format{
A data frame with 6728 rows and 6 variables:
\describe{
\item{time}{date and time of observation}
\item{X}{x position}
\item{Y}{y position}
\item{behaviour}{observed behaviour}
\item{grid}{grid of observation}
\item{zone}{zone of observation}
}
}
\source{
James Brereton
}
\usage{
skink_grid
}
\description{
A dataset containing skink data
}
\keyword{datasets}
