% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skink_raw}
\alias{skink_raw}
\title{Zoo monitor data for skink}
\format{
A data frame with 15201 rows and 4 variables:
\describe{
\item{DateTime}{date and time of observation}
\item{Space Use Coordinate X}{x position}
\item{Space Use Coordinate Y}{y position}
\item{Interval Channel 1 Value}{observed behaviour}
}
}
\source{
James Brereton
}
\usage{
skink_raw
}
\description{
A dataset containing skink data
}
\keyword{datasets}
