% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_species_names}
\alias{split_species_names}
\title{splits species names}
\usage{
split_species_names(x, model = "BirdNET v2.4")
}
\arguments{
\item{x}{string}

\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}
}
\description{
Split BirdNET species name in scientific and common name. If model perch v2 is used, try to lookup common names from BirdNET species list.
}
\keyword{internal}
