% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_device_id}
\alias{strip_device_id}
\title{Strip device used as prefix id from file names otherwise representing just a time-stamp.}
\usage{
strip_device_id(
  input_dir,
  pattern = ".WAV",
  recursive = T,
  ignore.case = T,
  .simulate = F
)
}
\arguments{
\item{input_dir}{input folder}

\item{pattern}{file pattern}

\item{recursive}{boolean}

\item{ignore.case}{boolean}

\item{.simulate}{boolean}
}
\description{
Retrieves timestamps from file names, expecting format YYYMMDD_HHMMSS as the sub-fix.
}
\details{
Use with caution as the functions simply retrieves the last 14 chacters of the file name. If unsure set \code{.simulate = TRUE}. Folders named 'extracted' are skipped!
}
