% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan_surfer.R
\name{swan_surfR}
\alias{swan_surfR}
\title{Reads in external data, combines with internal datasets and produces Swan
River surfer plots}
\usage{
swan_surfR(path)
}
\arguments{
\item{path}{Character string filepath to location of sonde data xlsx
workbooks.}
}
\value{
Two pdf format four panel surfer plots of the Swan River. One shows
   the full extent of the monitoring run (river mouth to the just beyond the
   site POL - Upper Swan Power Lines). The second shows from the Narrows
   Bridge to site POL.
}
\description{
\code{swan_surfR} takes a file path to Swan River sonde output and creates
    a four panel (single column) surfer plot of salinity, dissolved oxygen,
    chlorophyll a, and temperature in pdf format.
}
\details{
Surfer plots display a
    cross-section of the river where the metrics of interest have been
    interpolated between sonde locations. Inverse distance weighting has
    been used for the interpolation.

A river "bottom" is displayed which puts the interpolation in context.
    The river bottom has been derived from a combination of historical
    maximum depths at sampling locations as well as station points extracted
    from a "best" navigation line over the latest corporate bathymetry. See
    project documentation for further details.
}
\examples{
\dontrun{
swan_surfR(path = "Z:/DEC/MonitoringProgram/Data")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
