\name{test.equal.tagged}
\alias{test.equal.tagged}
\title{Specific Methods for Testing Equality}
\description{See the generic function description.}
\usage{
    \method{test.equal}{tagged}(obj1, obj2)
}
\arguments{
    \item{obj1}{object to be compared with obj2.}
    \item{obj2}{object to be compared with obj1.}
}
\seealso{ \code{\link{test.equal}} }
%\keyword{DSE}
\keyword{ts}

