\name{testEqual.tagged}
\alias{testEqual.tagged}
\title{Specific Methods for Testing Equality}
\description{See the generic function description.}
\usage{
    \method{testEqual}{tagged}(obj1, obj2)
}
\arguments{
    \item{obj1}{object to be compared with obj2.}
    \item{obj2}{object to be compared with obj1.}
}
\seealso{ \code{\link[tframe]{testEqual}} }
%\keyword{DSE}
\keyword{ts}

