\name{tfplot.combined.forecast}
\alias{tfplot.combined.forecast}
\title{Specific Methods for tfplot}
\description{See the generic function description.}
\usage{
    \method{tfplot}{combined.forecast}(combined.forecast,verbose=F, 
       start.=start(combined.forecast$data$output),
       Title="Projection", select.inputs=NULL, select.outputs=NULL,
       mar=par()$mar)
}
\seealso{ \code{\link{tfplot}} }
%\keyword{DSE}
\keyword{ts}

