\name{TSmeta.TSdata}
\alias{TSmeta,TSdata,missing-method}
\alias{TSmeta,TSestModel,missing-method}

\title{TSmeta Methods}
\description{See the generic function description.}
\usage{
    \S4method{TSmeta}{TSdata,missing}(x, con, ...)
    \S4method{TSmeta}{TSestModel,missing}(x, con, ...)
}
\arguments{
    \item{x}{An object of class TSdata or TSestModel.}
    \item{con}{A database connection.}
    \item{...}{Additional arguments ignored.}
}

\seealso{ \code{\link[TSdbi]{TSmeta}} }

\concept{DSE}
\keyword{ts}

