% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{wav2video}
\alias{wav2video}
\title{Generate video showing spectrogram}
\usage{
wav2video(
  wav,
  out = sub("\\\\.wav$", ".mp4", wav, ignore.case = TRUE),
  start = 0,
  dur = NULL,
  fps = 25,
  window_s = 10,
  fmax = 10,
  wl = 2048,
  ovlp = 75,
  dyn_range = 40,
  width = 1280,
  height = 540,
  pal = colorRampPalette(c("#000000", "#000123", "#000537", "#051586", "#0e9276",
    "#1dd752", "#1afe36"), bias = 1.6)(256),
  ffmpeg = "ffmpeg",
  keep_frames = FALSE,
  cores = 1
)
}
\arguments{
\item{wav}{wav file}

\item{out}{output mp4 file}

\item{start}{start within recording in seconds. defaults to 0}

\item{dur}{stop within recording in seconds. defaults to NULL}

\item{fps}{frames per second. defaults to 25}

\item{window_s}{visible window length in seconds. defaults to 10}

\item{fmax}{max. Frequency in kHz. defaults to 10}

\item{wl}{FFT-window. defaults to 1024}

\item{ovlp}{overlap}

\item{dyn_range}{dynamic range in dB}

\item{width}{width}

\item{height}{height}

\item{pal}{color palette}

\item{ffmpeg}{ffmpeg}

\item{keep_frames}{logical}

\item{cores}{number of cores}
}
\description{
Generate video showing spectrogram
}
