% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{write_birdnet_slist}
\alias{write_birdnet_slist}
\title{Write custom species list}
\usage{
write_birdnet_slist(
  sci_names = c("Bubo bubo", "Buteo butep"),
  filename = NULL,
  language = "de"
)
}
\arguments{
\item{sci_names}{character vector of scientific species names}

\item{filename}{name of text file (e.g. 'my_folder/custom_species_list.txt')}

\item{language}{language. defaults to 'de'}
}
\description{
Write custom species list
}
\examples{
# NOT RUN
## Read species list of ornitho.de
# sci_names <- readxl::read_xlsx('~/Artenlisten/Avifauna.xlsx', 'OrnithoV3', 'D7:D798', 'Names')
## Write species list to file
# out <- write_birdnet_slist(names[['Names']], "inst/extdata/ornitho_de.txt")

}
