% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdnet2audacity}
\alias{birdnet2audacity}
\title{Export birdnetR predictions to audacity label file}
\usage{
birdnet2audacity(predictions, wave_files)
}
\arguments{
\item{predictions}{list of data frames returned by \link[birdnetR]{predict_species_from_audio_file }}

\item{wave_files}{path to wave_files in order of data frames}
}
\description{
Creates a text file with labels for use in Audacity matching --rtype 'audacity' when using [BirdNET-Analyzer](https://github.com/birdnet-team/BirdNET-Analyzer)
}
