% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdNET_graph}
\alias{birdNET_graph}
\title{Daily activity graph}
\usage{
birdNET_graph(path, taxon, model = c("BirdNET_V2.4", "Perch v2"))
}
\arguments{
\item{path}{path}

\item{taxon}{Taxon name}

\item{model}{model used to classifiy}
}
\description{
Daily activity graph
}
