% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{birdNET_heatmap}
\alias{birdNET_heatmap}
\title{Heatmap of BirdNET detections}
\usage{
birdNET_heatmap(db, taxon)
}
\arguments{
\item{db}{data frame or path to xlsx file}

\item{taxon}{Taxon name (matched against the \code{Taxon} column)}
}
\description{
Heatmap of BirdNET detections
}
