% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdNET_process}
\alias{birdNET_process}
\title{Execute \link[=predict.birdnet_model_acoustic]{birdnet.predict()} with custom settings}
\usage{
birdNET_process(
  audio,
  model = c("BirdNET v2.4", "Perch v2"),
  slist = NULL,
  language = "de",
  batch_size = NULL,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25
)
}
\arguments{
\item{audio}{audio file}

\item{model}{one of c('BirdNET v2.4', 'Perch v2')}

\item{slist}{see \code{species_list} in \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{language}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{batch_size}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{min_confidence}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{chunk_overlap_s}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{sigmoid_sensitivity}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}
}
\value{
see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}
}
\description{
Runs \link[=predict.birdnet_model_acoustic]{birdnet.predict()} on audio files
}
