% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdNET_process}
\alias{birdNET_process}
\title{Execute \link[birdnetR]{predict_species_from_audio_file } with custom settings}
\usage{
birdNET_process(
  audio,
  slist = NULL,
  language = "de",
  batch_size = 1,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25
)
}
\arguments{
\item{audio}{audio file}

\item{slist}{species_list as character vector}

\item{language}{birdnetR::predict_species_from_audio_file}

\item{batch_size}{see \link[birdnetR]{predict_species_from_audio_file }}

\item{min_confidence}{see \link[birdnetR]{predict_species_from_audio_file }}

\item{chunk_overlap_s}{see \link[birdnetR]{predict_species_from_audio_file }}

\item{sigmoid_sensitivity}{see \link[birdnetR]{predict_species_from_audio_file }}
}
\value{
see \link[birdnetR]{predict_species_from_audio_file }
}
\description{
Execute \link[birdnetR]{predict_species_from_audio_file } with custom settings
}
