% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdNET_process_batch}
\alias{birdNET_process_batch}
\title{Split execution of \link{birdNET_process} in sets of wave files}
\usage{
birdNET_process_batch(
  wave_files,
  model = c("BirdNET v2.4", "Perch v2"),
  n = 12,
  slist = NULL,
  language = "de",
  batch_size = 3,
  restart = NULL,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25
)
}
\arguments{
\item{wave_files}{audio files to process}

\item{model}{one of c('BirdNET v2.4', 'Perch v2')}

\item{n}{number of audio files to process in a single run}

\item{slist}{see \code{species_list} in \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{language}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{batch_size}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{restart}{optional. restart iteration from here}

\item{min_confidence}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{chunk_overlap_s}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{sigmoid_sensitivity}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}
}
\description{
Split execution of \link{birdNET_process} in sets of wave files
}
