% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{birdNET_process_batch}
\alias{birdNET_process_batch}
\title{Split execution of \link{birdNET_process} in sets of wave files}
\usage{
birdNET_process_batch(
  wave_files,
  n = 12,
  slist = NULL,
  language = "de",
  batch_size = 3,
  restart = NULL,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25
)
}
\arguments{
\item{wave_files}{audio files to process}

\item{n}{number of audio files to process in a single run}

\item{slist}{species_list as character vector}

\item{language}{birdnetR::predict_species_from_audio_file}

\item{batch_size}{see \link[birdnetR]{predict_species_from_audio_file }}

\item{restart}{optional. restart iteration from here}

\item{min_confidence}{see \link[birdnetR]{predict_species_from_audio_file }}

\item{chunk_overlap_s}{see \link[birdnetR]{predict_species_from_audio_file }}

\item{sigmoid_sensitivity}{see \link[birdnetR]{predict_species_from_audio_file }}
}
\description{
Split execution of \link{birdNET_process} in sets of wave files
}
