% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alarms_constant_var_faster.R
\name{count_alarms_constant_var_faster}
\alias{count_alarms_constant_var_faster}
\title{Count and plot alarms using a spline, assuming constant variance}
\usage{
count_alarms_constant_var_faster(load_training, vib_training, knots, load_test,
  vib_test, show_figure = TRUE, load_min = 0)
}
\arguments{
\item{load_training}{A vector containing the training data of the load variables such as power or generator speed.}

\item{vib_training}{A vector containing the training data of the vibration signal.}

\item{knots}{A vector containing the locations of the knots.}

\item{load_test}{A vector with the test data of the load variable.}

\item{vib_test}{A vector with the test data of the vibration signal.}

\item{show_figure}{A logical vector indicating if a plot should be made.}

\item{load_min}{The minimum load to be used.}
}
\value{
A count of the number of alarms, alarm rate, and residuals. Furthermore a plot is
generated by default set by \code{show_figure}.
\item{count}{A count of the number of alarms in the test set.}
\item{alarm_rate}{The alarm rate in the test set.}
\item{residuals}{The residuals from the test set.}
}
\description{
The function uses a spline to fit to \code{vib_training} against \code{load_training}.
An alarm is generates when the test data fall outside a band of three sigma.
The function generates a plot by default, showing the training data falling outside the 3 sigma band.
}
\examples{
count_alarms_constant_var_faster(ltrain, vtrain, knots = c(20,25,27),
                                 ltest, vtest, load_min = 17)
}

