% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alarms_constant_var_faster.R
\name{count_alarms_constant_var_faster}
\alias{count_alarms_constant_var_faster}
\title{Count and plot alarms using a spline}
\usage{
count_alarms_constant_var_faster(load_training, vib_training, knots = NULL,
  load_test, vib_test, load_min = 0)
}
\arguments{
\item{load_min}{}
}
\description{
Count and plot alarms using a spline
}

