% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alarms_hetero_faster.R
\name{count_alarms_hetero_faster}
\alias{count_alarms_hetero_faster}
\title{Count and plot alarms using a spline with heterogeneous variance.}
\usage{
count_alarms_hetero_faster(load_training, vib_training, knots, load_test,
  vib_test, show_figure = TRUE, show_summary = FALSE, bins = 160,
  load_min = 0)
}
\arguments{
\item{load_training}{A vector containing the training data of the load variables such as power or generator speed.}

\item{vib_training}{A vector containing the Vibration signal.}

\item{knots}{A vector containing the locations of the knots.}

\item{load_test}{A vector with the test data.}

\item{vib_test}{A vector with the test data.}

\item{show_figure}{A logical vector indicating if a plot should be made.}

\item{show_summary}{A logical vector stating whether the quantile summaries should be shown or not.}

\item{bins}{The number of bins used.}

\item{load_min}{The minimum load used.}
}
\value{
A count of the number of alarms, alarm rate, and residuals. Further a plot is
generated if \code{show_figure = TRUE}.
}
\description{
Count and plot alarms using a spline with heterogeneous variance.
}
\examples{
count_alarms_hetero_var_faster(ltrain, vtrain, knots = c(20,25,27), ltest, vtest, load_min = 17)
}

