% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alarms_hetero_faster.R
\name{count_alarms_hetero_faster}
\alias{count_alarms_hetero_faster}
\title{Count and plot alarms using a spline with heterogeneous variance.}
\usage{
count_alarms_hetero_faster(load_training, vib_training, knots, load_test,
  vib_test, show_figure = TRUE, show_summary = FALSE, bins = 160,
  load_min = 0)
}
\arguments{
\item{load_min}{}
}
\description{
Count and plot alarms using a spline with heterogeneous variance.
}

