% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_WTs_func.R
\name{identify_WTs_func}
\alias{identify_WTs_func}
\title{Identifies the wind turbines if several unit IDs appear.}
\usage{
identify_WTs_func(dataframe, id = names(dataframe[1]))
}
\arguments{
\item{dataframe}{A data frame containing the data from condition monitoring. The data frame
should contain a column with unit IDs.}

\item{id}{The name as a character of the column with unit IDs.}
}
\value{
The data in \code{dataframe} sorted by \code{id} into a list containing data frames,
one for each wind turbines, and the time stamps are converted to POSIXct class. A string is
printed showing the number of wind turbines in \code{dataframe}.
}
\description{
Identifies the wind turbines if there are several unit IDs, and returns the same data in
a list containing data frames for each unit ID, and the time stamps are converted to POXIXct class.
}
\examples{
identify_WTs_func(data, id = "UnitID")
}

