% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_random_walk.R
\name{Kalman_filter_random_walk}
\alias{Kalman_filter_random_walk}
\title{Kalman filter applied to multivariate random walk.}
\usage{
Kalman_filter_random_walk(ts, F = c(1, 1, 1, 1), R = 0.1 * diag(4),
  Q = 0.1, m0 = 0, C0 = 1)
}
\arguments{
\item{ts}{A matrix containing the data from a multivariate time series.}

\item{R}{The covariance of the measurement noise.}

\item{Q}{The covariance of the state noise.}

\item{m0}{Initial state.}

\item{C0}{Initial covariance of the state process.}
}
\value{
Returns a list of the state values and covariances.
\item{state_values}{The state values.}
\item{state_cov}{The state covariances.}
}
\description{
Function that applies the Kalman filter to four observations assuming the
state is a univariate random walk. The observation and state equations
are \eqn{\bm{y}_t = Fx_t + \bm{v}_t} and \eqn{x_t = x_{t-1} + w_t}, respectively.
}

