\name{monitor-package}
\alias{monitor-package}
\alias{monitor}
\docType{package}
\title{
\packageTitle{monitor}
}
\description{
\packageDescription{monitor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{monitor}
\packageIndices{monitor}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{monitor}

Maintainer: \packageMaintainer{monitor}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
