% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{read_birdnet_slist}
\alias{read_birdnet_slist}
\title{Read and subset species list from models}
\usage{
read_birdnet_slist(
  model = c("BirdNET v2.4", "Perch v2"),
  language = "de",
  .cached = F
)
}
\arguments{
\item{model}{One of \code{c('BirdNET v2.4', 'Perch v2')}}

\item{language}{language. defaults to 'de'}

\item{.cached}{logical}
}
\description{
Read and subset species list from models
}
