% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{read_birdnet_slist}
\alias{read_birdnet_slist}
\title{Read and subset species list from BirdNET v2.4 model}
\usage{
read_birdnet_slist(language = "de")
}
\arguments{
\item{language}{language. defaults to 'de'}
}
\description{
Read and subset species list from BirdNET v2.4 model
}
