% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{run_birdnet}
\alias{run_birdnet}
\title{Split execution of \link{birdNET_process} in sets of wave files}
\usage{
run_birdnet(
  wave_files,
  model = c("BirdNET v2.4", "Perch v2"),
  slist = NULL,
  language = "de",
  batch_size = 3,
  min_confidence = 0.7,
  chunk_overlap_s = 0,
  sigmoid_sensitivity = 1.25,
  skip.existing.results = FALSE
)
}
\arguments{
\item{wave_files}{audio files to process}

\item{model}{one of c('BirdNET v2.4', 'Perch v2')}

\item{slist}{see \code{species_list} in \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{language}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{batch_size}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{min_confidence}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{chunk_overlap_s}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{sigmoid_sensitivity}{see \link[=predict.birdnet_model_acoustic]{birdnet.predict()}}

\item{skip.existing.results}{boolean}
}
\description{
Replaces \link{birdNET_process_batch}
}
