% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_waves}
\alias{split_waves}
\title{Split wave files in 10 minute segments}
\usage{
split_waves(path, pattern = c(".wav", ".mp3"))
}
\arguments{
\item{path}{path}

\item{pattern}{file format, .wav or .mp3}
}
\description{
Split wave files in 10 minute segments
}
