\name{testEqual.tagged}
\alias{testEqual.tagged}
\title{Specific Methods for Testing Equality}
\description{See the generic function description.}
\usage{
    \method{testEqual}{tagged}(obj1, obj2, fuzz= 1e-16)
}
\arguments{
    \item{obj1}{object to be compared with obj2.}
    \item{obj2}{object to be compared with obj1.}
    \item{fuzz}{tolerance for numerical comparisons.}
}
\seealso{ \code{\link[tframe]{testEqual}} }
\concept{DSE}
\keyword{ts}

