\name{tfwindow.tagged}
\alias{tfwindow.tagged}
\alias{tbind.tagged}
\alias{trim.na.tagged}
\alias{splice.tagged}
\alias{select.series.tagged}

\title{Specific Methods for tframed Data}
\description{See the generic function description.}
\usage{
	tfwindow(x)
	tbind(x)
	trim.na(x)
	splice(x)
	select.series(x)
}

\seealso{
	\code{\link{tfwindow}},
	\code{\link{tbind}},
	\code{\link{trim.na}}
	\code{\link{splice}}
	\code{\link{select.series}}
}

\keyword{DSE}
\keyword{ts}

